Ajuste Roles

+++ app/Http/Controllers/api/ClienteController.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers\api;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Cliente;
+use App\Models\User;
+use App\Models\Role;
+use Illuminate\Support\Facades\Auth;
 
 class ClienteController extends Controller
 {
@@ -14,8 +17,15 @@ class ClienteController extends Controller
     public function index()
     {
         //
-        $clientes = Cliente::with('facturas')->get();
-        return response()->json($clientes);
+        $user = Auth::user();
+
+        if ($user->role->nombre == 'admin') {
+            $clientes = Cliente::with('facturas')->get();
+            return response()->json($clientes);
+        } else if ($user->role->nombre == 'cliente') {
+            $clientes = Cliente::with('facturas')->where('numero_documento',$user->cliente->numero_documento)->get();
+            return response()->json($clientes);
+        }
     }
 
     /**
@@ -24,8 +34,22 @@ class ClienteController extends Controller
     public function store(Request $request)
     {
         //
-        $cliente = Cliente::create($request->all());
-        return response()->json($cliente, 201);
+        $user = Auth::user();
+        if ($user->role->nombre == 'admin') {
+            $role = Role::where('nombre', 'cliente')->first();
+            $user = User::create([
+                'name' => $request->name,
+                'email' => $request->email,
+                'password' => bcrypt($request->password),
+                'role_id' => $role->id,
+            ]);
+            $request->merge(["user_id"=>$user->id]);
+            //return response()->json($request, 201);
+            $cliente = Cliente::create($request->all());
+            return response()->json($cliente, 201);
+        } else {
+            return response()->json(['message' => 'No tiene permisos para crear un Cliente'], 403);
+        }
     }
 
     /**
@@ -34,11 +58,21 @@ class ClienteController extends Controller
     public function show(string $id)
     {
         //
-        $cliente = Cliente::find($id);
-        if ($cliente) {
-            return response()->json($cliente, 200);
-        } else {
-            return response()->json(['message' => 'Cliente not found'], 404);
+        $user = Auth::user();
+        if ($user->role->nombre == 'admin') {
+            $cliente = Cliente::find($id);
+            if ($cliente) {
+                return response()->json($cliente, 200);
+            } else {
+                return response()->json(['message' => 'Cliente not found'], 404);
+            }
+        } else if ($user->role->nombre == 'cliente') {
+            $cliente = Cliente::find($id);
+            if ($cliente && $cliente->numero_documento==$user->cliente->numero_documento) {
+                return response()->json($cliente, 200);
+            } else {
+                return response()->json(['message' => 'Cliente not found'], 404);
+            }
         }
     }
 
@@ -48,13 +82,24 @@ class ClienteController extends Controller
     public function update(Request $request, string $id)
     {
         //
-        $cliente = Cliente::find($id);
-        if ($cliente) {
-            $cliente->update($request->all());
-            return response()->json($cliente, 200);
-        } else {
-            return response()->json(['message' => 'Cliente not found'], 404);
-        }
+        $user = Auth::user();
+        if ($user->role->nombre == 'admin') {
+            $cliente = Cliente::find($id);
+            if ($cliente) {
+                $cliente->update($request->all());
+                return response()->json($cliente, 200);
+            } else {
+                return response()->json(['message' => 'Cliente not found'], 404);
+            }
+        } else if ($user->role->nombre == 'cliente') {
+            $cliente = Cliente::find($id);
+            if ($cliente && $cliente->numero_documento==$user->cliente->numero_documento) {
+                $cliente->update($request->all());
+                return response()->json($cliente, 200);
+            } else {
+                return response()->json(['message' => 'Cliente not found'], 404);
+            }
+        } 
     }
 
     /**
@@ -63,12 +108,17 @@ class ClienteController extends Controller
     public function destroy(string $id)
     {
         //
-        $cliente = Cliente::find($id);
-        if ($cliente) {
-            $cliente->delete();
-            return response()->json(['message' => 'Cliente deleted'], 200);
+        $user = Auth::user();
+        if ($user->role->nombre == 'admin') {
+            $cliente = Cliente::find($id);
+            if ($cliente) {
+                $cliente->delete();
+                return response()->json(['message' => 'Cliente deleted'], 200);
+            } else {
+                return response()->json(['message' => 'Cliente not found'], 404);
+            }
         } else {
-            return response()->json(['message' => 'Cliente not found'], 404);
+            return response()->json(['message' => 'No tiene permisos para eliminar un Cliente'], 403);
         }
     }
 }
